requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.4.3
python-docx==1.2.0
scikit-learn==1.8.0
scipy==1.17.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
Werkzeug==3.1.7
